+++ cli.py
@@ -1,10 +1,10 @@
 import typer
 
-from ai_code_testing import functional_testing as ft
-from metric_measurement.enum import CodeDataset, TextMetric
-from metric_measurement import textual_metrics as tm
-from classifiers.enum import Classifier
-from classifiers.run_classifiers import run_full_classification
+from src.metric_test_proxy.ai_code_testing import functional_testing as ft
+from src.metric_test_proxy.metric_measurement.enum import CodeDataset, TextMetric
+from src.metric_test_proxy.metric_measurement import textual_metrics as tm
+from src.metric_test_proxy.classifiers.enum import Classifier
+from src.metric_test_proxy.classifiers.run_classifiers import run_full_classification
 
 
 dataset_help = ('Dataset to test. `original` contains duplicate scripts. `distinct` is similar to `original` but with no'
