+++ src/metric_test_proxy/metric_measurement/enum.py
